requirements.txt
ansible==5.3.0
jmespath==0.10.0
python-vagrant==0.5.15
pytest==7.0.1
pytest-testinfra==6.6.0
molecule==3.6.1
molecule-vagrant==1.0.0
